BIM: migrate ArchStructure's make* functions to Arch module
